# New evalprocs for pointer indexing and range checking