--- a/lape.lpr
+++ b/lape.lpr
@@ -9,7 +9,7 @@ uses
   Interfaces, // this includes the LCL widgetset
   Forms, Main,
 
-  lpparser, lpcompiler, lptypes, lpeval, lpinterpreter, lpinterpreter_types,
+  lpparser, lpcompiler, lptypes, lpeval, lpeval_extra, lpinterpreter, lpinterpreter_types,
   lpmessages, lpvartypes, lpcodeemitter, lptree, lpdisassembler,
   lpvartypes_array, lpvartypes_ord, lpvartypes_record, lpvartypes_helper,
   lpinternalmethods, lpinternalmethods_algorithm,