--- a/lpinterpreter.pas
+++ b/lpinterpreter.pas
@@ -535,26 +535,6 @@ var
     Inc(Code, ocSize);
   end;
 
-  procedure DoDynArrayRangeCheck; {$IFDEF Lape_Inline}inline;{$ENDIF}
-  var
-    Arr: Pointer;
-    Len, Index: SizeInt;
-  begin
-    Dec(FStackPos, SizeOf(Pointer) + SizeOf(SizeInt));
-    Inc(Code, ocSize);
-
-    Arr := PPointer(@FStack[FStackPos])^;
-    Index := PSizeInt(@FStack[FStackPos + SizeOf(SizeInt)])^;
-
-    if (Arr <> nil) then
-      Len := PSizeInt(Arr)[-1] {$IFDEF FPC}+1{$ENDIF}
-    else
-      Len := 0;
-
-    if (Index < 0) or (Index >= Len) then
-      LapeExceptionFmt(lpeIndexOutOfRange, [Index, 0, Len - 1]);
-  end;
-
   procedure DoInitStackLen; {$IFDEF Lape_Inline}inline;{$ENDIF}
   var
     InitStackSize: TStackOffset;