--- a/lpinterpreter_opcodecase.inc
+++ b/lpinterpreter_opcodecase.inc
@@ -16,7 +16,6 @@ case opCodeP(Code)^ of
   ocGetCallerAddress: DoGetCallerAddress();
   ocGetScriptMethodName: DoGetScriptMethodName();
   ocDumpCallStack: DoDumpCallStack();
-  ocDynArrayRangeCheck: DoDynArrayRangeCheck();
   ocInitStackLen: DoInitStackLen();
   ocInitStack: DoInitStack();
   ocGrowStack: DoGrowStack();